Swapped out the add functions for a to_elmt()
It seemed odd to me the way the add functions were threading the counts
and XMLElement description mutably through all the add functions. I saw
no reason for it's extra passing of mutable values.

To me it makes more sense for the "add" functions to just convert the
Entity to an XMLElement and then have the calling (main) handle what to
do with it, such as adding to another XMLElement and incrementing a
count. This provides better encapsulation, and possible re-use later.

ideally I would have impl From<Entity> for XMLElement, but since
XMLElement, Entity, and From trait are not defined by me the orphan rule
comes into play. For now I just created my own ToElemt Trait and used
that, but there might be a cleaner way to do this. I felt like the
boilerplate of creating and shuffling to a newtype didn't make sense
here, but maybe it does, and I'll revisit this.