Added in a ToElmt trait to convert DXF Entities
This cleans up a bit of the code. I will probably create a struct at
some point to represent the elmt XML and have that create the XMLEntity
that way I can also implement the std From Trait in std from the
Entities.

Also updated the polyline and lwpolyline to use the closed status from
dxf crate to correctly output the closed state in the generate elmt.
this closes Issue #9