Finalized the storage QElement Structs
I'm pretty happy with this. The results come out pretty much identical
to the old streaming logic. This should make referring back to other
parts of the drawings (such as blocks) easier.

Now to clean up the old streaming logic I had side by side with this for
testing purposes, and take my test code with block support and add it in