Initial Support for Blocks
There are some bugs. In one my test DXF files the blocks overlap in the
resulting elemt file, so the coordinate positions aren't being used
correctly. I'm assuming I'm using the coordinates within the block for
the elements, but they need to be offset by the coordinates of the
insert element.

Also one of my blocks has a text object which didn't show up.