Adding in calculating the Block Offset
This commit adds in support for calculating the offset of the block
elements based on the insertion point of the block....However in my
current sample dxf, these values are always coming back as (0,0).
I'm not sure if this problems lies int he dxf crate, my sample file, or
what. I need to do a bit more digging.