Fixed Block location / offset
I was using the block placement coordinates which were within the block.
What I wanted to use was the position of the Insert Entity to offset the
graphic elements.

I need to figure out why the text in my block isn't coming through, and
then I think this is ready for more thorough testing before marking the
block support complete and tag a new release