Some Preliminary to support Scaling
This is some early stage work to support trying to scale the entities
in the DXF drawing before exporting the final element file. Many DXF
drawings come in quite small depending on the coordinates used etc, when
converted to pixels used by QET. With truncating the floats to 2 decimal
places this can also cause some inaccuracies in the conversion.