Forgot that the y directions are reversed
I forgot that bewtween dxf and QET elements the y directions are
reversed, so I needed to negate my y value. That's why they were so far
off. That being said there still seems to be a small issue. My y values
are still of by a value of 0.0729166666666667 (in my main sample file
for testing this)..which when scaled comes out to being off by 7px in
the y axis....what have I overlooked.