Preliminary work to scale text size to correct Pt
Added in some logic to convert between dxf text size in real world units
to font points. Swapped out the tuple TryFrom logic for the Objects enum
and DynamicText object for a Builder object. The tuple worked well to
get thigs up and running quickly, but needing to add in the units to pt
text scaling started to make it more cumbersome.

There is still plenty of more work to do for the font handling, but it's
a start.