Fixed blown out font scaling
Before I got he font to point conversion logic in, I was just scaling
the font size by the same scaling factor as the rest of the drawing
temporarily. Since the font is already scaled to points based on the
units, it doesn't need to be scaled again with the rest of the drawing.
The couple of drawings I was doing my main testing on were in mm, so
when it multiplied it by 2 again to convert mm to px, it wasn't as
drastic as a difference as a drawing in inches, where it was multiplying
it by a factor of 50.8. The text is no longer scaled by the drawing
scale factor as it's scaled separately as part of the conversion to pts.