Cleaning up some more Clippy Lints
Some of this is commenting out code. For example the enum ItemType has
several variants as used in QET, such as Project, Diagram etc. Initially
I put them in, partially because I wasn't sure all what was needed. For
example Project probably isn't, but what about ElementsContainer. Right
now nothing other than Element is being used, and perhaps later I will
delete this code, and hard code the string "element" for type when the
XML is written out. For now I'll leave it in, but comment it out to
suppress the Clippy lints.