Changed Scale back to having an x and a y factor
ABout 2 months ago in commit 39e424d I changed scaling from having an x
and y factor to a single factor. I couldn't think of any reason why we
would need to scale an entity by a skewed factor. Well now I know why.
When a block is inserted in a DXF it can be scaled by seperate x and y
factors. This needs to be handled for the drawing to convert correctly.

This changes the scaling factor back to haveing x and y, but also adds in support for the block scaling values.