Enhanced support for Deepley nested blocks
My previous attempt wouldn't recurse very well. If you had a single
level of nested blocks (as in the original test file) it worked great.
But once you got more levels of nesting it didn't work. Another test
file I was pointed at essentially was made up of 1 block, that nested
down to 6 or 7 layers of blocks within blocks.

This code actually recurses down and should get all of the deeply nested
blocsk. There are still a few issues, in the orginal test file with only
1 nested level I can see that some (but not all) of the objects are
offset from the position they should be in. I must not be carrying an
offset properly, but I wanted to get this checked in.