Added Tracing Crate
Added the tracing crate for some loggin and troubleshooting. I'm having
some odd issues with the deeply nested blocks I need to get sorted. I've
been meaning to add in some logging for a while and it will hopefully
help me track down these issues. Right now it's crude and uses the
defaults which dumps the logs to stdout (or is it stderr?), which will
cause issues actually running it from QET. For now while troubleshooting
it gets the job done, and I'll clean it up.