Logging Cleanup
The Logging was previously dumping to stdout while I was trying to
troubleshoot the issues with the deeply nested blocks. That won't work
with QET which is expecting the Elemnt XML file to be written to stdout
This change customizes the logging a bit and writes it to a file, as
well as adds basic support for changing the log level based on an ENV
var.

From my quick tests this looks like it's working, and I'll get it
committed, but I need to do more testing, and a bit more clean up
right now it should just be dumping the log in whatever folder dxf2elmt
was launched from, which might not always be ideal.