I deeply nested blocks now seem to work!
After a lot of headache it appears that I finally discovered what was my
issue with deeply nested blocks. I was considering the offset/insertion
point, but not the blocks base point. For the SVS drawing the block all
had a (0,0) base point so it didn't matter, but for my Phoenix Contact
test drawing that had 7 layers of deeply nested blocks this was not
always the case.

I need to test this over more drawings.

I was also so busy testing this branch of drawings of blocks that I made
a critical error on my object descent logic where I wasn't inserting the
any top elements. If it was a block it was fine because it would then
insert the sub elements, but if the drawing had no blocks, or was a mix,
stuff would get left out.