Fix a crash if it couldn't create the log file
I accidentally let slip through some bad log file creation into the last
release. Creation of the log file was being unwrapped. This came up
when Laurent tested it on his Mac and the directory it was in for being
called from Qet didn't have write permissions, so creating the log file
failed, and panic, halting the program without converting the dxf.

This change adds some proper error handling, and tries to dump the logs
to a file, but also to stderr. If the log file fails, it still dumps to
stderr