Fixed the logic for LwPolyline to Rect
I had made a couple of mistakes that were causing issues converting an
LWPolyline to a Rectangle Element.

The right_bound and top_bound elements were calling min_by instead of
max_by, returning the wrong values. When converting it to a rectangle
I forgot to invert the Y axis.

I probably need to do a bit more testing, but right now this is looking
good, and could possibly be ready to be merged.