Moved Shoelace Calculation to Arity Trait
The logic for the shoelace calculation of the polygon area was
duplicated between the circularity and rectangularity traits. I've
isolated the logic for this and moved it into the Arity trait as a
default implementation. This removes some duplicated code, but still
allows for a specific type to implement a different area function if
needed.

I've also renamed the is_rectangle function to is_rectangular, to match
the is_circular function name.